Add module to recognize git's wildmatch patterns
The function elvis_gitignore:wildmatch/2 accepts a glob pattern as used
in Git's ignore files (see man page gitignore(5) for more details), a
text string, and returns whether the string is contained in the set
described by the pattern.

The code in this first commit is a bit messy, but passes the numerous
unit tests we have.